+++ resources/misc_check.rb
@@ -3,7 +3,7 @@ property :misc_timeout,     Integer
 property :warmup,           Integer
 property :misc_dynamic,     [true, false]
 property :config_directory, String, default: '/etc/keepalived/checks.d'
-property :config_file,      String, default: lazy { ::File.join(config_directory, "keepalived_misc_check__port-#{name.to_s.gsub(/\s+/, '-')}__.conf") }
+property :config_file,      String, default: lazy { ::File.join(config_directory, "keepalived_misc_check__#{name.to_s.gsub(/\s+/, '-')}__.conf") }
 property :cookbook,         String, default: 'keepalived'
 property :source,           String, default: 'misc_check.conf.erb'